+++ app/controllers/settings_controller.rb
@@ -2,4 +2,6 @@ class SettingsController < ApplicationController
   def index; end
 
   def external_services; end
+
+  def notifications; end
 end